Feature ETP-81: Minor fixes to drawer
